Develop to main (#52)
* Account for appearance BBox.

Customer found Flattening an annotation appearance resulted in a misplaced output.

Normally we use the annotation's rectangle to set the postional offset of our concatenate matrix, but in this case the BBox of the appearance form is a bit offset from its own 'origin'.

So we end up duplicating the positional offset and it gets thrown a good ways away from the expected location.

Check for this scenario and just force the BBox to start at 0.

* Fix transformation of annotation position.

We recently updated this sample based on a particular file and then the cusotmer followed up with another one so we need further adjustments.

* Add a link for Documentation and re-label the existing link API Reference

* Clean up leaks in sample.

---------

Co-authored-by: Arnulfo Arroyo <aarroyo@datalogics.com>
Co-authored-by: Richard Gailiunas <107433294+datalogics-rgailiunas@users.noreply.github.com>
Co-authored-by: Robert Boehne <robb@datalogics.com>
Co-authored-by: datalogics-saharay <39740722+datalogics-saharay@users.noreply.github.com>