Add support for setting the background color during Rendering.
Occasionally there's a desire for setting the background color and it's not obvious based on the API how to accomplish this as the most typical case is just a white background.

For more advanced usage, it's desirable to set the value of the alpha channel for rendering to DeviceRGBA/DeviceCMYKA which are special colorspace types PDFL recognizes for rendering
to a buffer that includes an Alpha Channel which is placed in the Image's SoftMask.

Add support for this to the sample with commentary about how to use it.

This is disabled by default in the sample as it's not the typical use case.

Also, some sample cleanup.